requirement.txt
streamlit==1.35.0
pandas==2.2.2
scikit-surprise==1.1.4
